Rewriting StatTracker class to extend Silex\Application

+++ src/BlueHerons/StatTracker/OCR.php
@@ -44,7 +44,9 @@ class OCR {
 			$resp->status = $thing;
 		}
 
-		echo json_encode($resp);
+                $json = json_encode($resp);
+                self::logger()->debug(sprintf("Sending payload: %s", $json));
+                echo $json;
 		ob_flush();
 		flush();
 	}
@@ -56,14 +58,14 @@ class OCR {
 	 *
 	 * @return array array of stat key's and values from the screenshot
 	 */
-	public static function scanAgentProfile($imagePath) {
+	public static function scanAgentProfile($imagePath, $stats) {
 		try {
 			$response = new StdClass();
 			$response->status = array();
 			self::logger()->debug(sprintf("Beginning scan of %s", $imagePath));
 			$imagePath = self::convertToPBM($imagePath);
 			$lines = self::executeOCR($imagePath);
-			$data = self::processAgentData($lines);
+			$data = self::processAgentData($lines, $stats);
 			self::logger()->debug("Parsed Stats:", $data);
 			$response->status = "Your screenshot has been processed.<p/>Please review your stats and click the \"Submit Stats\" button to submit.";
 			$response->stats = $data;
@@ -256,7 +258,7 @@ class OCR {
 		}
 		catch (Exception $e) {
 			copy($imagePath, $imagePath . "_errored");
-			unlink($newFile);
+			if (file_exists($newFile)) unlink($newFile);
 			throw $e;
 		}
 		finally {
@@ -299,7 +301,7 @@ class OCR {
 	 *
 	 * @return path to new PBM image
 	 */
-	public static function processAgentData($lines) {
+	public static function processAgentData($lines, $stats) {
 		try {
 			self::sendMessage("Processing scanned results...");
 			$step = 'start';
@@ -371,7 +373,6 @@ class OCR {
 			$elements = preg_replace('/g/', '9', $elements);
 
 			$data = array();
-			$stats = StatTracker::getStats();
 
 			foreach ($stats as $stat) {
 				if ($stat->ocr) {
